+++ lib/controllers/payment_option.rb
