+++ lib/controllers/product.rb
